requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.8.1
    # via
    #   aiohttp-retry
    #   fsspec
aiohttp-retry==2.4.6
    # via dvc
aiosignal==1.2.0
    # via aiohttp
appdirs==1.4.4
    # via dvc
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
asyncssh==2.10.1
    # via scmrepo
atpublic==3.0.1
    # via flufl-lock
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.0
    # via nbconvert
certifi==2021.10.8
    # via
    #   dulwich
    #   requests
cffi==1.15.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
    #   pygit2
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
colorama==0.4.4
    # via
    #   dvc
    #   ipython
    #   tqdm
commonmark==0.9.1
    # via rich
configobj==5.0.6
    # via dvc
cryptography==37.0.1
    # via asyncssh
cycler==0.11.0
    # via matplotlib
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
dictdiffer==0.9.0
    # via dvc
diskcache==5.4.0
    # via dvc
distro==1.7.0
    # via dvc
dpath==2.0.6
    # via dvc
dulwich==0.20.35
    # via scmrepo
dvc==2.10.2
    # via -r requirements.in
dvc-render==0.0.5
    # via
    #   dvc
    #   dvclive
dvclive==0.7.3
    # via dvc
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fastjsonschema==2.15.3
    # via nbformat
flatten-dict==0.4.2
    # via dvc
flufl-lock==7.0
    # via dvc
fonttools==4.33.3
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.3.0
    # via
    #   dvc
    #   scmrepo
ftfy==6.1.1
    # via python-benedict
funcy==1.17
    # via
    #   dvc
    #   dvc-render
    #   scmrepo
future==0.18.2
    # via grandalf
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via scmrepo
grandalf==0.6
    # via dvc
idna==3.3
    # via
    #   requests
    #   yarl
ipykernel==6.13.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.3.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
    #   qtconsole
ipywidgets==7.7.0
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via scikit-learn
jsonschema==4.4.0
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.3.0
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.3
    # via jupyter
jupyter-core==4.10.0
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.0
    # via ipywidgets
kiwisolver==1.4.2
    # via matplotlib
mailchecker==4.1.16
    # via python-benedict
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.5.2
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
nanotime==0.5.2
    # via dvc
nbclient==0.6.0
    # via nbconvert
nbconvert==6.5.0
    # via
    #   jupyter
    #   notebook
nbformat==5.3.0
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
networkx==2.8
    # via dvc
notebook==6.4.11
    # via
    #   -r requirements.in
    #   jupyter
    #   widgetsnbextension
numpy==1.22.3
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
packaging==21.3
    # via
    #   dvc
    #   ipykernel
    #   matplotlib
    #   nbconvert
    #   qtpy
pandas==1.4.2
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via
    #   dvc
    #   scmrepo
phonenumbers==8.12.47
    # via python-benedict
pickleshare==0.7.5
    # via ipython
pillow==9.1.0
    # via matplotlib
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.29
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.0
    # via
    #   dvc
    #   flufl-lock
    #   ipykernel
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydot==1.4.2
    # via dvc
pygit2==1.9.1
    # via scmrepo
pygments==2.12.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
    #   rich
pygtrie==2.4.2
    # via
    #   dvc
    #   scmrepo
pyparsing==3.0.8
    # via
    #   dvc
    #   grandalf
    #   matplotlib
    #   packaging
    #   pydot
pyrsistent==0.18.1
    # via jsonschema
python-benedict==0.25.1
    # via dvc
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   python-benedict
python-fsutil==0.6.0
    # via python-benedict
python-slugify==6.1.2
    # via python-benedict
pytz==2022.1
    # via pandas
pywin32==303
    # via jupyter-core
pywinpty==2.0.5
    # via terminado
pyyaml==6.0
    # via python-benedict
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.0
    # via jupyter
qtpy==2.0.1
    # via qtconsole
requests==2.27.1
    # via
    #   dvc
    #   fsspec
    #   python-benedict
rich==12.3.0
    # via dvc
ruamel-yaml==0.17.21
    # via dvc
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
scikit-learn==1.1.1
    # via -r requirements.in
scipy==1.8.1
    # via
    #   -r requirements.in
    #   scikit-learn
scmrepo==0.0.19
    # via dvc
send2trash==1.8.0
    # via notebook
shortuuid==1.0.8
    # via dvc
shtab==1.5.4
    # via dvc
six==1.16.0
    # via
    #   bleach
    #   configobj
    #   flatten-dict
    #   python-dateutil
smmap==5.0.0
    # via gitdb
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.2.0
    # via ipython
tabulate==0.8.9
    # via
    #   dvc
    #   dvc-render
terminado==0.13.3
    # via notebook
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.1.0
    # via scikit-learn
tinycss2==1.1.1
    # via nbconvert
toml==0.10.2
    # via
    #   dvc
    #   python-benedict
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.64.0
    # via dvc
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==4.2.0
    # via
    #   asyncssh
    #   dvc
urllib3==1.26.9
    # via
    #   dulwich
    #   requests
voluptuous==0.13.1
    # via dvc
wcwidth==0.2.5
    # via
    #   ftfy
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==3.6.0
    # via ipywidgets
xmltodict==0.12.0
    # via python-benedict
yarl==1.7.2
    # via aiohttp
zc-lockfile==2.0
    # via dvc

# The following packages are considered to be unsafe in a requirements file:
# setuptools
